requirements.txt
aiohttp
alembic
beautifulsoup4
chroma-hnswlib
chromadb
crewai
crewai-tools
embedchain
fastapi
GitPython
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-storage
gptcache
huggingface-hub
instructor
kubernetes
lancedb
langchain
langchain-community
langchain-openai
langchain-text-splitters
langchain-together
langchainhub
lxml
numpy
openai
pandas
peewee
posthog
pulsar-client
pydantic
pydantic-settings
PyGithub
pypdf
pytest
python-dotenv
python-multipart
pytube
pyyaml
requests
selenium
SQLAlchemy
tiktoken
together
tqdm
uvicorn
yfinance
youtube-transcript-api
yt-dlp
